worked on cli

+++ mmu_sim_cmd.h
@@ -14,7 +14,7 @@
  * @param vaddr the virtual address of the byte to be read
  * @return the byte stored at the specified virtual address
  */
-uint16_t mmu_sim_read(addr_t vaddr);
+uint8_t mmu_sim_read(addr_t vaddr);
 
 /**
  * Reads the specified number of bytes starting at the specified virtual address.
@@ -22,14 +22,14 @@ uint16_t mmu_sim_read(addr_t vaddr);
  * @param nbytes the number of bytes to be read
  * @return the bytes stored starting at the specified starting address
  */
-uint16_t* mmu_sim_readn(addr_t vaddr, int nbytes);
+uint8_t* mmu_sim_readn(addr_t vaddr, int nbytes);
 
 /**
  * Writes the specified byte value at the specified virtual address.
  * @param vaddr the virtual address to write to
  * @param val the byte to be written
  */
-void mmu_sim_write(addr_t vaddr, uint16_t val);
+void mmu_sim_write(addr_t vaddr, uint8_t val);
 
 /**
  * Writes the specified word (2 byte) value starting at the specified virtual address.
@@ -37,7 +37,7 @@ void mmu_sim_write(addr_t vaddr, uint16_t val);
  * @param val1 the first byte of the word to be written
  * @param val2 the second byte of the word to be written
  */
-void mmu_sim_writew(addr_t vaddr, uint16_t val1, uint16_t val2);
+void mmu_sim_writew(addr_t vaddr, uint8_t val1, uint8_t val2);
 
 /**
  * Writes the specified double word (4 byte) value starting at the specified virtual address.
@@ -47,7 +47,7 @@ void mmu_sim_writew(addr_t vaddr, uint16_t val1, uint16_t val2);
  * @param val3 the third byte of the double word to be written
  * @param val4 the fourth byte of the double word to be written
  */
-void mmu_sim_writedw(addr_t vaddr, uint16_t val1, uint16_t val2, uint16_t val3, uint16_t val4);
+void mmu_sim_writedw(addr_t vaddr, uint8_t val1, uint8_t val2, uint8_t val3, uint8_t val4);
 
 /**
  * Writes a zero value for the specified number of bytes starting at the specified virtual address.
